requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Werkzeug==3.1.3
python-dotenv==1.0.1
stripe==12.0.0
email-validator==2.2.0
